Refactor schema cardinality types
Use a Map instead of an object for CardinalityRule so that inserts are
O(1) instead of O(n). Also, using a Set<CardinalityRule> was wrong,
only a CardinalityRule was needed; this refactor fixes that error.

Define Cardinality as an Enum so that we can refer to individual
cardinalities (useful for the default cardinality, for instance).