Add schema schema test
Make FileInfos.content lazy, so that we don't read files unless we need
to. This is not memoized, because keeping file contents in memory
in-between tests requires a lot of memory.

Allow parse() to take a schema so that we can instantiate Parser easily

Add better error messages when schema validation tests fail.