Add error classes
This allows us to assign an error code to each error (like TypeScript),
to have centralize error messages, and to provide information about the
errors to user.

For instance, once the parser encodes file positions in the AST (which
it currently doesn't), we can display the positions at which the schema
validation error happened.