Add schema definition errors
For clearer error messages, we can think about rule 3 (pairwise
intersection of prop content tags should be empty) differently: it is
equivalent to saying that each content tag should be assigned to at
most one prop in each tag. 

This allows us to show all the incorrect assignments in the same error 
message. With this re-statement of property 3, we do not need the 
`intersection` function anymore.