Add cardinality error checking
If we encode the cardinality constraints as min and max, we can check
cardinality constraints as we build the IR. The idea is to check that
we're not passing the max as we open blocks in the IRHandler, and to
ensure that we have the minimum required of all children when we close
a block. 

Not only is it easier to implement constraints this way, but this also
opens up the possibility of encoding custom constraints in the schema 
by specifying a min and max rather than one of the predefined 
cardinality constraints (oneOrMore, etc).

This also required a small change in the parser: because blocks are
checked when we close them, we must close all blocks when we're done
parsing. The stack is thus empty when we finish parsing, so we must
keep a reference to the root outside of the stack to be able to return
the result in `getResult()`.

This might be problematic for incremental compilation, but the
solution then would be to separate closing (popping from the stack)
from checking a block.