Redesign SchemaDefinition
The SchemaDefinition has been redesigned. This new design aims to 
enable better prop types. Note that this commit only changes the schema
side of things -- the parser still does not take the schema types into
account.

"Native types" (i.e. other than IRNode) are only available in the head
of a block tag, or in the args of an inline tag. This is an easier 
first step, and stems from the idea that a prop may have multiple tags
of the same type (i.e. multiple string tags), and that the consuming
code should be able to distinguish these. Therefore, while head and arg
props can contain a "native type", body props cannot.

This is also convenient, as body props are defined with indirection 
(i.e. referring to other tags, that may not have been transformed into
a `BlockSchema` yet). Keeping the aforementionned property ensures that
the props' types can be computed immediately; the computation doesn't 
need to be deferred until the root `Schema` object has been fully 
constructed. It also allows for circular references in the props 
without looping infinitely, a real concern if we delayed this 
computation.

Better discriminated union types in the SchemaDefinition (on `raw` and
`rawBody`) make the consuming code in Schema.ts easier to write.

`DuplicatePropTagsError` has been removed, as it is no longer needed in
the new design: the schema props are defined as objects with tag names
as keys, meaning that no duplicates tags can exist in a prop. 
Duplication can still exist between the head and the body, as well as
between inline arg names, so `DuplicatePropNameError` remains.

The schema generators have been rewritten to fit the new schema. The
test schemas have also been rewritten using the new schema generators.
As a nice side-effect confirming the new design, these schemas were
short and easy to write.

The root tag is no longer treated as a regular block tag. As a result,
the root no longer has a "head" in the IR, so all test resources were
changed accordingly. Additionally, the rewritten document schema no
longer ignores the head of `#code` block tags.