Remove Cardinality in favor of CardinalityConstraint
A CardinalityConstraint is more general, so this change actually adds
expressive power to the schemas. Removing Cardinality also allows us to
simplify the internal representation to a single type.

At the same time, we still maintain the ease of use of Cardinality by
providing generator functions with the same names.

To simplify, we also rename CardinalityConstraint to Cardinality.