Fix local pkg-ok checks by deleting dist folder before build
When introducing a change that should break the pkg-ok check, the check
would sometimes still pass because the dist folder contained artifacts 
from the previous build (which was valid).

Travis CI starts from a clean state and will therefore detect this
breaking change anyway, but this commit enables us to detect the 
breakage earlier, locally.

I considered both the rimraf and del libraries for compatibility with
Windows; rimraf is more popular and smaller, so I chose that.