Move Typescript from devDepencies to dependencies
Because of the Typescript code generation, which is a runtime usage of
the Typescript compiler API, we must declare Typescript as a dependency
instead of a devDependency. 

Down the line, we may want to write our own Typescript AST interfaces
and code printer in order to get rid of this dependency, and be able to
run on the Web, and not just Node.js.